added json build step to vite config as a plugin
