+++ .prettierignore
@@ -8,6 +8,7 @@ node_modules
 !.env.example
 
 src/langs/data/json/
+src/langs/r.json
 
 # Ignore files for PNPM, NPM and YARN
 pnpm-lock.yaml
