+++ .prettierrc
@@ -7,7 +7,6 @@
   "trailingComma": "none",
   "printWidth": 100,
   "plugins": ["prettier-plugin-svelte"],
-  "pluginSearchDirs": ["."],
   "overrides": [
     {
       "files": "*.svelte",
