+++ package.json
@@ -4,12 +4,12 @@
   "private": true,
   "scripts": {
     "dev": "vite dev --host",
-    "build": "tsx build.ts && vite build",
+    "build": "vite build",
     "preview": "vite preview",
     "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
     "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
-    "lint": "prettier --plugin-search-dir . --check .",
-    "format": "prettier --plugin-search-dir . --write .",
+    "format-lint": "prettier --check .",
+    "format": "prettier --write .",
     "test": "vitest",
     "run-test": "vitest run"
   },
@@ -25,6 +25,7 @@
     "@zerodevx/svelte-toast": "^0.9.5",
     "autoprefixer": "^10.4.19",
     "bcrypt": "^5.1.1",
+    "cross-env": "^7.0.3",
     "eslint": "^8.57.0",
     "js-yaml": "^4.1.0",
     "jsonwebtoken": "^9.0.2",
