+++ src/app.html
@@ -1,4 +1,4 @@
-<!DOCTYPE html>
+<!doctype html>
 <html lang="%lang%">
   <head>
     <meta charset="utf-8" />
