+++ vite.config.js
@@ -1,9 +1,9 @@
 import { sveltekit } from '@sveltejs/kit/vite';
-import path from 'path';
+import { lang_data_json_build_plugin } from './src/langs/json_build_plugin';
 
 /** @type {import('vite').UserConfig} */
 const config = {
-  plugins: [sveltekit()],
+  plugins: [sveltekit(), lang_data_json_build_plugin()],
   server: {
     port: 3427,
     strictPort: true
